# internal/api/v1/docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Tailnet API",
    "title": "Tailnet API",
    "termsOfService": "https://github.com/sudosu404/providers/blob/main/LICENSE",
    "contact": {
      "name": "Yusing",
      "url": "https://github.com/sudosu404/providers/issues"
    },
    "license": {
      "name": "MIT",
      "url": "https://github.com/sudosu404/providers/blob/main/LICENSE"
    },
    "version": "1.0"
  },
  "basePath": "/api/v1",
  "paths": {
...
    "/cert/info": {
      "get": {
        "description": "Get cert info",
        "produces": [
          "application/json"
        ],
        "tags": [
          "cert"
        ],
        "summary": "Get cert info",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/CertInfo"
            }
          },
          "403": {
            "description": "Forbidden",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        },
        "x-id": "info",
        "operationId": "info"
      }
    },
...
    "/docker/container/{id}": {
      "get": {
        "description": "Get container by container id",
        "produces": [
          "application/json"
        ],
        "tags": [
          "docker"
        ],
        "summary": "Get container",
        "parameters": [
          {
            "type": "string",
            "description": "Container ID",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ContainerResponse"
            }
          },
          "400": {
            "description": "ID is required",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "403": {
            "description": "Forbidden",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "404": {
            "description": "Container not found",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        },
        "x-id": "container",
        "operationId": "container"
      }
    },
...
    "/docker/restart": {
      "post": {
        "description": "Restart container by container id",
        "produces": [
          "application/json"
        ],
        "tags": [
          "docker"
        ],
        "summary": "Restart container",
        "parameters": [
          {
            "description": "Request",
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/dockerapi.StopRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/SuccessResponse"
            }
          },
          "400": {
            "description": "Invalid request",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "403": {
            "description": "Forbidden",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "404": {
            "description": "Container not found",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        },
        "x-id": "restart",
        "operationId": "restart"
      }
    },
...
    "/favicon": {
      "get": {
        "description": "Get favicon",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "image/svg+xml",
          "image/x-icon",
          "image/png",
          "image/webp"
        ],
        "tags": [
          "v1"
        ],
        "summary": "Get favicon",
        "parameters": [
          {
            "type": "string",
            "description": "URL of the route",
            "name": "url",
            "in": "query"
          },
          {
            "type": "string",
            "description": "Alias of the route",
            "name": "alias",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/homepage.FetchResult"
              }
            }
          },
          "400": {
            "description": "Bad Request: alias is empty or route is not HTTPRoute",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "403": {
            "description": "Forbidden: unauthorized",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "404": {
            "description": "Not Found: route or icon not found",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error: internal error",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        },
        "x-id": "favicon",
        "operationId": "favicon"
      }
    },
    "/file/content": {
      "get": {
        "description": "Get file content",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json",
          "application/godoxy+yaml"
        ],
        "tags": [
          "file"
        ],
        "summary": "Get file content",
        "parameters": [
          {
            "type": "string",
            "format": "filename",
            "name": "filename",
            "in": "query",
            "required": true
          },
          {
            "enum": [
              "config",
              "provider",
              "middleware"
            ],
            "type": "string",
            "x-enum-varnames": [
              "FileTypeConfig",
              "FileTypeProvider",
              "FileTypeMiddleware"
            ],
            "name": "type",
            "in": "query",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "string"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "403": {
            "description": "Forbidden",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        },
        "x-id": "get",
        "operationId": "get"
      },
      "put": {
        "description": "Set file content",
        "consumes": [
          "text/plain"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "file"
        ],
        "summary": "Set file content",
        "parameters": [
          {
            "enum": [
              "config",
              "provider",
              "middleware"
            ],
            "type": "string",
            "description": "Type",
            "name": "type",
            "in": "query",
            "required": true
          },
          {
            "type": "string",
            "description": "Filename",
            "name": "filename",
            "in": "query",
            "required": true
          },
          {
            "description": "File",
            "name": "file",
            "in": "body",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/SuccessResponse"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "403": {
            "description": "Forbidden",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        },
        "x-id": "set",
        "operationId": "set"
      }
    },
...
    "/icons": {
      "get": {
        "description": "List icons",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "v1"
        ],
        "summary": "List icons",
        "parameters": [
          {
            "type": "integer",
            "description": "Limit",
            "name": "limit",
            "in": "query"
          },
          {
            "type": "string",
            "description": "Keyword",
            "name": "keyword",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/homepage.IconMetaSearch"
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "403": {
            "description": "Forbidden",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        },
        "x-id": "icons",
        "operationId": "icons"
      }
    },
...
    "/route/providers": {
      "get": {
        "description": "List route providers",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "route",
          "websocket"
        ],
        "summary": "List route providers",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/RouteProvider"
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        },
        "x-id": "providers",
        "operationId": "providers"
      }
    },
...
  },
  "definitions": {
...
    "CertInfo": {
      "type": "object",
      "properties": {
        "dns_names": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "x-nullable": false,
          "x-omitempty": false
        },
        "email_addresses": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "x-nullable": false,
          "x-omitempty": false
        },
        "issuer": {
          "type": "string",
          "x-nullable": false,
          "x-omitempty": false
        },
        "not_after": {
          "type": "integer",
          "x-nullable": false,
          "x-omitempty": false
        },
        "not_before": {
          "type": "integer",
          "x-nullable": false,
          "x-omitempty": false
        },
        "subject": {
          "type": "string",
          "x-nullable": false,
          "x-omitempty": false
        }
      },
      "x-nullable": false,
      "x-omitempty": false
    },
...
    "ContainerResponse": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "x-nullable": false,
          "x-omitempty": false
        },
        "image": {
          "type": "string",
          "x-nullable": false,
          "x-omitempty": false
        },
        "name": {
          "type": "string",
          "x-nullable": false,
          "x-omitempty": false
        },
        "server": {
          "type": "string",
          "x-nullable": false,
          "x-omitempty": false
        },
        "state": {
          "allOf": [
            {
              "$ref": "#/definitions/ContainerState"
            }
          ],
          "x-nullable": true
        }
      },
      "x-nullable": false,
      "x-omitempty": false
    },
    "ContainerState": {
      "type": "string",
      "enum": [
        "created",
        "running",
        "paused",
        "restarting",
        "removing",
        "exited",
        "dead"
      ],
      "x-enum-comments": {
        "StateCreated": "StateCreated indicates the container is created, but not (yet) started.",
        "StateDead": "StateDead indicates that the container failed to be deleted. Containers in this state are attempted to be cleaned up when the daemon restarts.",
        "StateExited": "StateExited indicates that the container exited.",
        "StatePaused": "StatePaused indicates that the container's current state is paused.",
        "StateRemoving": "StateRemoving indicates that the container is being removed.",
        "StateRestarting": "StateRestarting indicates that the container is currently restarting.",
        "StateRunning": "StateRunning indicates that the container is running."
      },
      "x-enum-descriptions": [
        "StateCreated indicates the container is created, but not (yet) started.",
        "StateRunning indicates that the container is running.",
        "StatePaused indicates that the container's current state is paused.",
        "StateRestarting indicates that the container is currently restarting.",
        "StateRemoving indicates that the container is being removed.",
        "StateExited indicates that the container exited.",
        "StateDead indicates that the container failed to be deleted. Containers in this state are attempted to be cleaned up when the daemon restarts."
      ],
      "x-enum-varnames": [
        "StateCreated",
        "StateRunning",
        "StatePaused",
        "StateRestarting",
        "StateRemoving",
        "StateExited",
        "StateDead"
      ],
      "x-nullable": false,
      "x-omitempty": false
    },
...
    "ErrorResponse": {
      "type": "object",
      "properties": {
        "error": {
          "type": "string",
          "x-nullable": true
        },
        "message": {
          "type": "string",
          "x-nullable": false,
          "x-omitempty": false
        }
      },
      "x-nullable": false,
      "x-omitempty": false
    },
...
    "RouteProvider": {
      "type": "object",
      "properties": {
        "full_name": {
          "type": "string",
          "x-nullable": false,
          "x-omitempty": false
        },
        "short_name": {
          "type": "string",
          "x-nullable": false,
          "x-omitempty": false
        }
      },
      "x-nullable": false,
      "x-omitempty": false
    },
...
    "SuccessResponse": {
      "type": "object",
      "properties": {
        "details": {
          "type": "object",
          "additionalProperties": {},
          "x-nullable": true
        },
        "message": {
          "type": "string",
          "x-nullable": false,
          "x-omitempty": false
        }
      },
      "x-nullable": false,
      "x-omitempty": false
    },
...
    "dockerapi.StopRequest": {
      "type": "object",
      "required": [
        "id"
      ],
      "properties": {
        "id": {
          "type": "string",
          "x-nullable": false,
          "x-omitempty": false
        },
        "signal": {
          "description": "Signal (optional) is the signal to send to the container to (gracefully)\nstop it before forcibly terminating the container with SIGKILL after the\ntimeout expires. If not value is set, the default (SIGTERM) is used.",
          "type": "string"
        },
        "timeout": {
          "description": "Timeout (optional) is the timeout (in seconds) to wait for the container\nto stop gracefully before forcibly terminating it with SIGKILL.\n\n- Use nil to use the default timeout (10 seconds).\n- Use '-1' to wait indefinitely.\n- Use '0' to not wait for the container to exit gracefully, and\n  immediately proceeds to forcibly terminating the container.\n- Other positive values are used as timeout (in seconds).",
          "type": "integer"
        }
      },
      "x-nullable": false,
      "x-omitempty": false
    },
    "homepage.FetchResult": {
      "type": "object",
      "properties": {
        "icon": {
          "type": "array",
          "items": {
            "type": "integer",
            "format": "int32"
          },
          "x-nullable": false,
          "x-omitempty": false
        },
        "statusCode": {
          "type": "integer",
          "x-nullable": false,
          "x-omitempty": false
        }
      },
      "x-nullable": false,
      "x-omitempty": false
    },
    "homepage.IconMetaSearch": {
      "type": "object",
      "properties": {
        "Dark": {
          "type": "boolean",
          "x-nullable": false,
          "x-omitempty": false
        },
        "Light": {
          "type": "boolean",
          "x-nullable": false,
          "x-omitempty": false
        },
        "PNG": {
          "type": "boolean",
          "x-nullable": false,
          "x-omitempty": false
        },
        "Ref": {
          "type": "string",
          "x-nullable": false,
          "x-omitempty": false
        },
        "SVG": {
          "type": "boolean",
          "x-nullable": false,
          "x-omitempty": false
        },
        "Source": {
          "$ref": "#/definitions/homepage.IconSource",
          "x-nullable": false,
          "x-omitempty": false
        },
        "WebP": {
          "type": "boolean",
          "x-nullable": false,
          "x-omitempty": false
        }
      },
      "x-nullable": false,
      "x-omitempty": false
    },
    "homepage.IconSource": {
      "type": "string",
      "enum": [
        "https://",
        "@target",
        "@walkxcode",
        "@selfhst"
      ],
      "x-enum-varnames": [
        "IconSourceAbsolute",
        "IconSourceRelative",
        "IconSourceWalkXCode",
        "IconSourceSelfhSt"
      ],
      "x-nullable": false,
      "x-omitempty": false
    },
...
  },
...
}